move: vaccine 테이블 fetch 함수 통합?
